(GH-14) Transition from SCSS vars to CSS vars
Prior to this commit, all modifiable values in the style for a platen
site were set via SCSS variables; this is useful for override at
*build* time but does not allow for easy override at *run* time - ie,
with a mode/theme toggle interface.

This commit transitions all SCSS variables, except the name of the
game being overridden, to CSS variables.

The name must stay as an SCSS variable for the purposes of being
used as a selector instead of a property value.